% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinetransform.R
\name{arcsinetransform}
\alias{arcsinetransform}
\title{Arc sine transformation of the numeric vector}
\usage{
arcsinetransform(numeric.vector, type, n)
}
\arguments{
\item{numeric.vector}{data vector to be transformed}

\item{type}{0 if data is in percentage and 1 if data is in proportion}

\item{n}{is the number of units upon which the percentage/proportion data is based}
}
\value{
Arc sine transformed data
}
\description{
The function divide values by 100, does square root and than sin inverse of each values of vector.
    If any of the values of a vector is 0 or 100, it is replaced by 1/4n or 100-(1/4n), respectively.
}
\examples{
vector<-c(23,0,29.6,35.6,33,35.6,10.5,100)
# Arc sine trnasformation for percentage data and n=10
arcsinetransform(vector,0,10)
}
