% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert the data frame into list of numeric nature}
\usage{
convert(data1)
}
\arguments{
\item{data1}{data-frame to be converted into list}
}
\value{
list of numeric vectors
}
\description{
Convert the data frame into list of numeric nature
}
