% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logretransform.R
\name{logretransform}
\alias{logretransform}
\title{Re-transform the log transformed data}
\usage{
logretransform(transformed.mean, if.zero.present)
}
\arguments{
\item{transformed.mean}{vector of mean which needs to be re-transformed}

\item{if.zero.present}{0 if zero was present in the data prior to transformation of data.
1 if zero was absent in the data prior to transformation}
}
\value{
Log re-transformed values
}
\description{
Re-transform the log transformed data. When log transformation is done, the mean of the treatments needs to be re-transformed for comparison.
}
\examples{
vector<-c(0,2.004,1.114,1.491,1.431,1.415,1.845)
#Re-transformation of data with zero present in data prior to transformation
logretransform(vector,0)
}
