% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtransform.R
\name{logtransform}
\alias{logtransform}
\title{Log transformation of the numeric vector}
\usage{
logtransform(numeric.vector)
}
\arguments{
\item{numeric.vector}{data vector to be transformed}
}
\value{
A list of
\itemize{
  \item \code{Ratio}- A ratio of maximum and minimum values of the data
  \item \code{LogTransformedVector} - A vector of the transformed data
  \item \code{Comment} - A comment about zero being present in data or not
}
}
\description{
The function carries out log with base 10 transformation of each values of vector.
    If one of values of a vector is 0, 1 is added to each observation.
    Log transformation is carried out for the data when variance is proportional to square of the mean and treatment effects are multiplicative in nature.
}
\examples{
vector<-c(100,0,120,1000,52,30,60)
logtransform(vector)
}
