% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsddata.R
\docType{data}
\name{lsddata}
\alias{lsddata}
\title{Data for Latin Square Design}
\format{
The data has 5 columns and 25 rows
\describe{
\item{Row}{Consist sequence of rows. Row consists of 5 levels}
\item{Column}{Consist sequence of column. Column consists of 5 levels}
\item{Treatment}{Consist sequence of treatments. There are 5 treatments A, B, C, D & E}
\item{Yield}{Yield as dependent variable}
\item{Plant Height}{Plant height as dependent variable}
}
}
\usage{
lsddata
}
\description{
The data consists of Rows, Columns, Treatments and two dependent variables Yield and Plant Height. The data is generated manually.
}
\keyword{datasets}
