% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrttransform.R
\name{sqrttransform}
\alias{sqrttransform}
\title{Square root transformation of the numeric vector}
\usage{
sqrttransform(numeric.vector)
}
\arguments{
\item{numeric.vector}{data vector to be transformed}
}
\value{
Square root transformed data
}
\description{
The function carries out square root transformation of each values of vector.
    If one of values of a vector is 0, 0.5 is added to each observation.
}
\examples{
vector<-c(0,25,36,6,9,25,70)
sqrttransform(vector)
}
